cocotechoes_drive-2024-Question-8
</>cocotechoes drive 2024 

Question 8 : 
Given a list of students, each student has 5 subjects, and for each subject, we have, total
marks, and marks scored. Find the top five students. For example Amit Sharma (English, 100,
60 | Hindi, 100, 70 | Math 1, 70, 40 | Math 2, 70 60, Science, 100, 80)
Please note you need to create random data of about 20 students with names like “Student1”,
“Student2” and so on. Each student will have 5 subjects with random values for marks scored
ranging from 0 to total marks for the subject.

Hint: Use an array with an array or an array of map or an array of objects